Modified migration scrripts, added entity classes, and added controller service and DTOs to connect to external api

+++ src/main/resources/db/migration/V11052025152600__front_table.sql
@@ -1,6 +1,6 @@
 CREATE TABLE IF NOT EXISTS front(
-    front_id SERIAL PRIMARY KEY,
+    front_id BIGSERIAL PRIMARY KEY,
     content VARCHAR(2000) UNIQUE NOT NULL,
-    card_id INTEGER NOT NULL,
+    card_id BIGINT NOT NULL,
     CONSTRAINT front_card_fk FOREIGN KEY (card_id) REFERENCES cards(card_id) ON DELETE CASCADE
 );
