+++ src/main/resources/db/migration/V11052025152700__back_table.sql
@@ -1,6 +1,6 @@
 CREATE TABLE IF NOT EXISTS back(
-    back_id SERIAL PRIMARY KEY,
+    back_id BIGSERIAL PRIMARY KEY,
     content VARCHAR(2000) UNIQUE NOT NULL,
-    card_id INTEGER NOT NULL,
+    card_id BIGINT NOT NULL,
     CONSTRAINT back_card_fk FOREIGN KEY (card_id) REFERENCES cards(card_id) ON DELETE CASCADE
 );
